Removing warnings from compilation

+++ Marlin/src/backtrace/unwarmbytab.cpp
@@ -185,7 +185,7 @@ static UnwResult UnwTabExecuteInstructions(const UnwindCallbacks *cb, UnwTabStat
       /* pop r4-r[4+nnn] or pop r4-r[4+nnn], r14*/
       vsp = ucb->vrs[13];
 
-      for (reg = 4; reg <= (instruction & 0x07) + 4; ++reg) {
+      for (reg = 4; reg <= uint32_t((instruction & 0x07) + 4); ++reg) {
         uint32_t v;
         if (!cb->readW(vsp,&v))
           return UNWIND_DREAD_W_FAIL;
